Snapshot of SHAPE-dev stable
