+++ examples/ase_test.py
@@ -52,18 +52,30 @@ print ("VASP_stress:\n", atoms.get_stress())
 
 ###################################################################################################
 
-from ase.calculators.espresso import Espresso
 import kp_finder
+import qepy
+
+try:
+    from mpi4py import MPI
+    comm = MPI.COMM_WORLD
+except Exception:
+    comm = None
+
+from qepy.calculator import QEpyCalculator
 
 kpoints = kp_finder.get_kpoints(kgrid=0.07)
 
 pseudopotentials = {'Si': 'Si.pbe-n-rrkjus_psl.1.0.0.UPF'}
-path_to_pseudopotentials="$HOME/apps/SSSP_1.3.0_PBE_efficiency"
-command = 'mpirun -np 16 $HOME/apps/qe-7.2/bin pw.x -in PREFIX.pwi > PREFIX.pwo'
+# path_to_pseudopotentials="$HOME/apps/SSSP_1.3.0_PBE_efficiency"
+# command = 'mpirun -np 16 $HOME/apps/qe-7.2/bin pw.x -in PREFIX.pwi > PREFIX.pwo'
+try:
+    os.environ["ESPRESSO_PSEUDO"]
+except KeyError:
+    os.system("export ESPRESSO_PSEUDO=$HOME/apps/SSSP_1.3.0_PBE_efficiency")
 
 input_data = {
     'control': {
-        'calculation': 'scf',
+        'calculation': 'vc-relax',
         'prefix': 'silicon',
         'outdir': './',
         'etot_conv_thr': 1.0e-5,
@@ -79,15 +91,21 @@ input_data = {
         'nosym': True },
     'electrons': {
         'electron_maxstep': 800,
-        'diagonalization': 'rmm-davidson',
+        # 'diagonalization': 'rmm-davidson', # Not implemented in QEpy
         'mixing_mode': 'local-TF',
         'mixing_beta': 0.5,
         'conv_thr': 1.0e-6 }
 }
 
-calc1 = Espresso(input_data = input_data,
-                 pseudopotentials = pseudopotentials,
-                 kpts = tuple(kpoints))
+ase_espresso = {
+    'input_data': input_data,
+    'pseudopotentials': pseudopotentials,
+    'kpts': tuple(kpoints)
+}
+
+calc1 = QEpyCalculator(comm = comm,
+                       ase_espresso = ase_espresso,
+                       logfile='QE.log')
 
 atoms.calc = calc1
 print ("QE_energy:\n", atoms.get_potential_energy())
@@ -268,7 +286,7 @@ print ("M3GNet_stress:\n", atoms.get_stress())
 ###################################################################################################
 
 
-from fplib3_api4ase import fp_GD_Calculator
+from shape.calculator import SHAPE_Calculator
 from functools import reduce
 
 chem_nums = list(atoms.numbers)
@@ -276,7 +294,7 @@ znucl_list = reduce(lambda re, x: re+[x] if x not in re else re, chem_nums, [])
 ntyp = len(znucl_list)
 znucl = znucl_list
 
-calc1 = fp_GD_Calculator(
+calc1 = SHAPE_Calculator(
             cutoff = 4.0,
             contract = False,
             znucl = znucl,
