+++ setup.py
@@ -1,0 +1,25 @@
+from setuptools import setup, find_packages
+
+setup(
+    name="shape",
+    version="1.1.0",
+    packages=find_packages(),
+    install_requires=[
+        # List your dependencies here
+        # For example:
+        # 'numpy',
+        # 'scipy',
+    ],
+    author="Zhu Research Group",
+    author_email="li.zhu@rutgers.edu",
+    description="A brief description of the shape library",
+    long_description=open("README.md").read(),
+    long_description_content_type="text/markdown",
+    url="https://github.com/yourusername/shape",
+    classifiers=[
+        "Programming Language :: Python :: 3",
+        "License :: OSI Approved :: MIT License",
+        "Operating System :: OS Independent",
+    ],
+    python_requires='>=3.6',
+)
